[hotfix] dynamic error growThePie
